Correction erreur deconnexion et gestion si est connecté ou pas

+++ php/deconnexion.php
@@ -1,12 +1,24 @@
 <?php
     session_start();
-    $membres = unserialize(file_get_contents('../Membres.data'));
+    function estConnecte(){
+        if(isset($_SESSION["estConnecte"])){
+            return  $_SESSION["estConnecte"];
+        }
+        else{
+            return false;
+        }
+        return false;
+    }
+    
+    if(file_exists("Membres.data") && estConnecte()){
+        $membres = unserialize(file_get_contents('../Membres.data'));
+        
+        $membres[$_SESSION["InfoMembre"]["id"]]["panier"] = array_merge($membres[$_SESSION["InfoMembre"]["id"]]["panier"],$_SESSION["InfoMembre"]["panier"]);
+        $_SESSION["estConnecte"]=false;
 
-    $membres[$_SESSION["InfoMembre"]["id"]]["panier"] = array_merge($membres[$_SESSION["InfoMembre"]["id"]]["panier"],$_SESSION["InfoMembre"]["panier"]);
-    $_SESSION["estConnecte"]=false;
+        file_put_contents('Membres.data', serialize($membres));
 
-    file_put_contents('Membres.data', serialize($membres));
-
-    session_destroy();
-    echo '<script> window.location.href="../index.php" </script>' ;
+        session_destroy();
+        echo '<script> window.location.href="../index.php" </script>' ;
+    }
 ?>
